+++ packages/core/src/testing/index.ts
@@ -1,7 +1,6 @@
 import '@prosekit/pm/view/style/prosemirror.css'
 
 import type { Attrs } from '@prosekit/pm/model'
-import { userEvent } from '@vitest/browser/context'
 
 import { union } from '../editor/union'
 import { defineBaseCommands } from '../extensions/command'
@@ -181,8 +180,4 @@ export function setupTest() {
     m,
     n: { ...n, p: n.paragraph },
   }
-}
-
-export async function inputText(input: string): Promise<void> {
-  return await userEvent.keyboard(input)
 }
