+++ packages/core/src/testing/keyboard.ts
@@ -1,0 +1,5 @@
+import { userEvent } from '@vitest/browser/context'
+
+export async function inputText(input: string): Promise<void> {
+  return await userEvent.keyboard(input)
+}
