+++ packages/extensions/src/heading/heading-keymap.spec.ts
@@ -4,10 +4,8 @@ import {
   it,
 } from 'vitest'
 
-import {
-  pressKey,
-  setupTest,
-} from '../testing'
+import { setupTest } from '../testing'
+import { pressKey } from '../testing/keyboard'
 
 describe('defineHeadingKeymap', () => {
   it('should toggle heading', async () => {
