+++ packages/extensions/src/testing/index.ts
@@ -7,7 +7,6 @@ import {
   defineHistory,
   defineParagraph,
   defineText,
-  isApple,
   union,
   type Extension,
   type ExtractMarkActions,
@@ -18,7 +17,6 @@ import {
   createTestEditor,
   type TestEditor,
 } from '@prosekit/core/test'
-import { userEvent } from '@vitest/browser/context'
 
 import { defineBlockquote } from '../blockquote'
 import { defineBold } from '../bold'
@@ -126,37 +124,4 @@ export function setupTest() {
       expanded: listWithAttrs({ kind: 'toggle', collapsed: false }),
     },
   }
-}
-
-/**
- * @example
- *
- * ```ts
- * await pressKey('mod-1')
- * await pressKey('Backspace')
- * ```
- *
- * @internal
- */
-export async function pressKey(input: string) {
-  const keys = input.split('-').map((key) => {
-    if (key.toLowerCase() === 'mod') {
-      return isApple ? 'Meta' : 'Control'
-    }
-    return key
-  })
-  const seq: string[] = []
-  for (const key of keys) {
-    // Press key without releasing it
-    seq.push('{' + key + '>}')
-  }
-  for (const key of keys.toReversed()) {
-    // Release a previously pressed key
-    seq.push('{/' + key + '}')
-  }
-  return await userEvent.keyboard(seq.join(''))
-}
-
-export async function inputText(input: string) {
-  return await userEvent.keyboard(input)
 }
